+++ jquery.reel.js
@@ -66,9 +66,14 @@
       // [NEW] in version 1.1
       delay:             -1, // delay before autoplay in seconds (no autoplay by default)
       friction:         0.9, // friction of the inertial rotation (will loose 90% of speed per second)
+      image:      undefined, // image sprite to be used
+      images:            [], // sequence array of individual images to be used instead of sprite
       inertial:        true, // drag & throw will give the rotation a momentum when true
       monitor:    undefined, // stored value name to monitor in the upper left corner of the viewport
       rebound:          0.5,
+      loading: 'Loading...', // label used for preloader
+      path:              '', // URL path to be prepended to `image` or `images` filenames
+      preloader:       true, // whether display the preloader or not
       revolution: undefined, // distance mouse must be dragged for full revolution
       speed:              0, // animated rotation speed in revolutions per second (Hz)
       step:       undefined, // initial step (overrides `frame`)
@@ -80,14 +85,15 @@
 
   $.fn.reel= function(options){
     var
+      set= $.extend({}, defaults, options),
       applicable= (function(tags){
         // Only IMG tags with non-empty SRC and non-zero WIDTH and HEIGHT will pass
         var
           pass= []
-        tags.filter(_img_).each(function(ix){
+        tags.each(function(ix){
           var
             $this= $(this),
-            src= $this.attr(_src_),
+            src= set.images.length && set.images || set.image || $this.attr(_src_),
             width= number($this.css(_width_)),
             height= number($this.css(_height_))
           if (!src || src == __ || !width || !height) return;
@@ -98,7 +104,6 @@
         });
         return $(pass);
       })(this),
-      set= $.extend({}, defaults, options),
       instances= []
 
     ticker= ticker || (function run_ticker(){
@@ -131,7 +136,7 @@
               id= t.attr(_id_),
               classes= t.attr(_class_),
               styles= t.attr('style'),
-              image= src.replace(/^(.*)\.(jpg|jpeg|png|gif)$/, '$1' + set.suffix + '.$2'),
+              images= set.images,
               size= { x: number(t.css(_width_)), y: number(t.css(_height_)) },
               turntable= $(_div_tag_).attr(_class_, classes).addClass(klass).addClass(set.klass),
               image_css= touchy || !set.saves ? { display: 'none' } : { opacity: 0 }
@@ -139,10 +144,11 @@
             .parent().attr(_id_, id).bind(on).css({
               display: 'block',
               width: size.x + _px_,
-              height: size.y + _px_,
-              backgroundImage: url(image)
+              height: size.y + _px_
             }))[0]);
-            store(_frames_, set.frames);
+            store(_image_, images.length && images.length || set.image || src.replace(/^(.*)\.(jpg|jpeg|png|gif)$/, '$1' + set.suffix + '.$2'));
+            store(_frame_, set.frame);
+            store(_frames_, images.length || set.frames);
             store(_spacing_, set.spacing);
             store(_dimensions_, size);
             store(_fraction_, 0);
@@ -159,7 +165,7 @@
           },
           teardown: function(e){
             t= t.unbind(on)
-            .find(dot(indicator_klass) + ', ' + dot(monitor_klass)).remove().end()
+            .find([dot(indicator_klass), dot(monitor_klass), dot(preloader_klass), _img_+dot(preloaded_frame_klass)].join(', ')).remove().end()
             .find(_img_)
             .attr(t.data(_backup_)).unwrap()
             .bind('setup', function resetup(e){
@@ -176,7 +182,13 @@
               frames= recall(_frames_),
               resolution= max(frames, recall(_steps_)),
               fraction= store(_fraction_, 1 / resolution * ((set.step || set.frame) - 1)),
-              frame= store(_frame_, fraction * frames + 1)
+              frame= store(_frame_, fraction * frames + 1),
+              image= recall(_image_),
+              images= set.images,
+              loaded= 0,
+              loading= set.loading,
+              preload= !images.length ? [image] : images,
+              $preloader
             if (!touchy) hotspot
               .css({ cursor: 'ew-resize' })
               .mouseenter(function(e){ t.trigger('enter') })
@@ -224,6 +236,10 @@
               className: monitor_klass,
               css: { position: _absolute_, left: 0, top: 0 }
             }));
+            set.preloader && t.append($preloader= $(_div_tag_, {
+              className: preloader_klass,
+              css: { position: _absolute_, right: 0, top: 0 }
+            }).text(loading+___+loaded+'/'+preload.length));
             set.indicator && t.append($(_div_tag_)
               .addClass(indicator_klass)
               .css({
@@ -233,6 +249,15 @@
                 position: _absolute_,
                 backgroundColor: _hex_black_
               }));
+            // Preloading of image(s)
+            $.each(preload, function preload_image(ix, url){
+              $(tag(_img_), { src: set.path+url, className: preloaded_frame_klass }).appendTo(t).hide().load(function image_loaded(){
+                $(this).remove();
+                loaded++;
+                $preloader.text(loading+___+loaded+'/'+preload.length);
+                loaded == preload.length && $preloader.remove();
+              });
+            });
             t.trigger('frameChange');
           },
           animate: function(e){
@@ -335,6 +360,8 @@
               fraction= !frame ? recall(_fraction_) : store(_fraction_, round_to(6, (frame-1) / (frames-1))),
               frame= !frame ? recall(_frame_) : store(_frame_, frame),
               frame= store(_frame_, round(frame)),
+              image= recall(_image_),
+              images= set.images,
               space= recall(_dimensions_),
               steps= recall(_steps_),
               spacing= recall(_spacing_),
@@ -355,7 +382,7 @@
                 reverse_shift= rows * major_size + (rows - 1) * spacing,
                 x= reversed && set.horizontal ? x - reverse_shift : x,
                 y= reversed && !set.horizontal ? y - reverse_shift : y,
-                shift= set.horizontal ? [y + _px_, x + _px_] : [x + _px_, y + _px_]
+                shift= images.length ? [0, 0] : set.horizontal ? [y + _px_, x + _px_] : [x + _px_, y + _px_]
             }else{
               var
                 travel= set.loops ? set.stitched : set.stitched - space.x,
@@ -364,9 +391,11 @@
                 shift= [-x + _px_, y + _px_]
             }
             var
+              sprite= images[frame - 1] || image,
               travel= space.x - set.indicator,
-              indicator= min_max(0, travel, round(fraction * (travel+2)) - 1)
-            t.css({ backgroundPosition: shift.join(' ') })
+              indicator= min_max(0, travel, round(fraction * (travel+2)) - 1),
+              css= { background: url(set.path+sprite)+___+shift.join(___) }
+            t.css(css)
               .find(dot(indicator_klass)).css({ left: indicator + _px_ });
           }
         },
@@ -396,6 +425,8 @@
     ns= '.reel',
     klass= 'jquery-reel',
     indicator_klass= 'indicator',
+    preloader_klass= 'preloader',
+    preloaded_frame_klass= 'preloaded_frame',
     monitor_klass= 'monitor',
     tick_event= 'tick'+ns,
     pool= $(document),
@@ -410,14 +441,15 @@
     // Storage keys
     _backup_= 'backup', _clicked_= 'clicked', _clicked_location_= 'clicked_location',
     _clicked_on_= 'clicked_on', _dimensions_= 'dimensions', _fraction_= 'fraction', _frame_= 'frame',
-    _frames_= 'frames', _last_fraction_= 'last_fraction', _reversed_= 'reversed', _spacing_= 'spacing',
-    _steps_= 'steps', _velocity_= 'velocity',
+    _frames_= 'frames', _image_= 'image', _last_fraction_= 'last_fraction', _reversed_= 'reversed',
+    _spacing_= 'spacing', _steps_= 'steps', _velocity_= 'velocity',
 
     // Various string primitives
-    __= '', _absolute_= 'absolute', _class_= 'class', _div_= 'div', _div_tag_= '<'+ _div_ + '/>',
+    __= '', ___= ' ', _absolute_= 'absolute', _class_= 'class', _div_= 'div', _div_tag_= tag(_div_),
     _height_= 'height', _hex_black_= '#000', _id_= 'id', _img_= 'img', _px_= 'px', _src_= 'src',
     _title_= 'title', _width_= 'width'
 
+  function tag(string){ return '<' + string + '/>' }
   function dot(string){ return '.' + string }
   function url(location){ return 'url(' + location + ')' }
   function round_to(decimals, number){ return +number.toFixed(decimals) }